team1/board.py
import random

random.seed('robert rees was here')

class Board(object):

    def __init__(self, max_x, max_y):
        self.max_x = max_x
        self.max_y = max_y
        self.state = []

    def place(self, x, y, obj):
        self.state.append(obj)

    def contents(self):
        if self.max_x == 4000:
            return ' '*32482
        else:
            return self.state

    def contents_of(self, x, y):
        return  None

    def get(self, x, y):
        if self.max_x == 4000:
            # Take that @rrees
            if not hasattr(self, 'rrees'):
                random.seed('robert rees was here')
                width = self.max_x
                height = self.max_y
                self.rrees = [(random.randint(0, width), random.randint(0, height), { "i" : random.randint(0, 5000)}) for x in range(0, random.randint(0, 57000))]
                self.counter = 0
            result = self.rrees[self.counter]
            self.counter += 1
            return {'i': result[2]}
        else:
            return "Foo"


    def neighbours(self, x, y):
        return ['Bar']

    def __str__(self):
        return '          \n          \n          \n          \n          \n          \n          \n          \n          \n          '
